+++ back/Core/Services/FFmpeg/FFmpeg.cs
@@ -1,4 +1,5 @@
 ﻿using System.Diagnostics;
+using System.Text;
 using Core.Enums;
 using Core.Utils;
 
@@ -6,12 +7,14 @@ namespace Core.Services.FFmpeg;
 
 public class FFmpeg
 {
+    private readonly string ffmpegPath = Env.Get("FFMPEG_PATH", @"C:\Program Files\FFmpeg\bin\ffmpeg.exe")!;
     private readonly string input;
 
-    private readonly string output = Env.Get("OUTPUT_FILE_PATH", Path.GetTempFileName() + ".mp4")!;
-    private readonly Process proc;
+    private readonly string output;
 
-    private readonly string ffmpegPath = Env.Get("FFMPEG_PATH", @"C:\Program Files\FFmpeg\bin\ffmpeg.exe");
+
+    private readonly string outputFolder = Env.Get("OUTPUT_FOLDER_PATH", Path.Join(Path.GetTempPath(), "video-share"))!;
+    private readonly Process proc;
 
 
     public Action<double> OnProgress = d => { };
@@ -19,25 +22,26 @@ public class FFmpeg
     public FFmpeg(string input, VideoFormat format)
     {
         this.input = input;
-
+        output = Path.Join(outputFolder, "output.mp4");
+        EnsureOutputFolder();
         var formatStr = format switch
         {
             VideoFormat.Streamable => "h264_nvenc",
             _ => throw new ArgumentOutOfRangeException(nameof(format), $"Not expected format value: {format}")
         };
 
+
         proc = new Process
         {
             StartInfo = new ProcessStartInfo
             {
                 FileName = ffmpegPath,
-                Arguments = $"-i \"{input}\" -c:v \"{formatStr}\" -vf subtitles=\"{Path.GetFileName(input)}\"  -pix_fmt yuv420p -y \"{output}\" ",
-                WorkingDirectory = Path.GetDirectoryName(output),
+                Arguments = $"-i \"{input}\" -c:v \"{formatStr}\"   -pix_fmt yuv420p -y \"{output}\" ",
+                WorkingDirectory = Path.GetDirectoryName(input),
                 CreateNoWindow = true,
                 RedirectStandardError = true,
                 UseShellExecute = false,
                 WindowStyle = ProcessWindowStyle.Hidden
-
             }
         };
 
@@ -52,16 +56,20 @@ public class FFmpeg
     {
         var nbFrames = await GetFramesCount();
 
+        var stderr = new StringBuilder();
+
         proc.ErrorDataReceived += (proc, e) =>
         {
-            Console.WriteLine("Error " +  e.Data?.ToString());
             if (e.Data == null) return;
-            if (e.Data.Contains("frame="))
+            stderr.AppendLine(e.Data.ToString());
+            Console.WriteLine("Error " + e.Data?.ToString());
+
+            if (e.Data!.Contains("frame="))
             {
                 var currentFrame = GetCurrentFrame(e.Data);
-                var percentage = currentFrame / (double)nbFrames * 100;
+                var percentage = currentFrame / (double) nbFrames * 100;
 
-                if(percentage > 100) percentage = 100;
+                if (percentage > 100) percentage = 100;
 
                 OnProgress(percentage);
             }
@@ -72,13 +80,9 @@ public class FFmpeg
         proc.BeginErrorReadLine();
         await proc.WaitForExitAsync();
 
-        if (proc.ExitCode != 0)
-        {
-            var stderr = await proc.StandardError.ReadToEndAsync();
-            throw new Exception("FFmpeg error: " + stderr);
-        }
+        if (proc.ExitCode != 0) throw new Exception("FFmpeg error: " + stderr);
 
-        var content =  await File.ReadAllBytesAsync(output);
+        var content = await File.ReadAllBytesAsync(output);
         return content;
     }
 
@@ -88,10 +92,10 @@ public class FFmpeg
         var properties = await FFprobe.GetFileInfos(input);
         var stream = properties.Streams.Find(stream => stream.CodecType == "video")!;
         var parts = stream.AvgFrameRate.Split("/");
-        var framerate = (long)Math.Round(double.Parse(parts[0]) / double.Parse(parts[1]));
+        var framerate = (long) Math.Round(double.Parse(parts[0]) / double.Parse(parts[1]));
 
         var replace = properties.Format.Duration.Replace(".", ",");
-        return (long)Math.Round(framerate * double.Parse(replace));
+        return (long) Math.Round(framerate * double.Parse(replace));
     }
 
 
@@ -99,15 +103,18 @@ public class FFmpeg
     {
         var data = chunk.Split("\n").Select(line => line.Split("=")).ToArray();
 
-         return long.Parse(data[0][1].Trim().Split(" ")[0]);
+        return long.Parse(data[0][1].Trim().Split(" ")[0]);
     }
 
     public void Clean()
     {
-        if(File.Exists(output))
-        {
-            File.Delete(output);
-        }
+        if (File.Exists(output)) File.Delete(output);
+    }
+
+
+    private void EnsureOutputFolder()
+    {
+        Directory.CreateDirectory(outputFolder);
     }
 }
 // ffmpeg  -i '.\Mushoku Tensei - 01.mkv' -c:v h264_nvenc -vf subtitles='Mushoku Tensei - 01.mkv'  -pix_fmt yuvj420p -y '.\Mushoku Tensei - 01.mp4'
